backend/file_processor/pyproject.toml
[tool.poetry]
name = "file_processor"
version = "0.1.0"
description = ""
authors = ["ilyakochankov <ilyakochankov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
dynaconf = "^3.1.8"
"pdfminer.six" = "^20220524"
pandas = "^1.4.2"
numpy = "^1.22.4"
python-snappy = "^0.6.1"
faust-streaming = "^0.8.5"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
